requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements/requirements.txt requirements/requirements.in
#
ansi2html==1.9.2
    # via -r requirements/requirements.in
anyio==4.12.1
    # via httpx
asgiref==3.11.0
    # via
    #   django
    #   django-browser-reload
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
automat==25.4.16
    # via twisted
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.3.0
    # via
    #   parsel
    #   scrapy
defusedxml==0.7.1
    # via scrapy
django==5.2.10
    # via
    #   -r requirements/requirements.in
    #   django-appconf
    #   django-autocomplete-light
    #   django-browser-reload
    #   django-debug-toolbar
    #   django-extensions
    #   djangorestframework
django-admin-interface==0.32.0
    # via -r requirements/requirements.in
django-appconf==1.2.0
    # via django-imagekit
django-autocomplete-light==3.12.1
    # via -r requirements/requirements.in
django-browser-reload==1.21.0
    # via -r requirements/requirements.in
django-cleanup==9.0.0
    # via -r requirements/requirements.in
django-colorfield==0.14.0
    # via django-admin-interface
django-debug-toolbar==6.1.0
    # via -r requirements/requirements.in
django-environ==0.12.0
    # via -r requirements/requirements.in
django-extensions==4.1
    # via -r requirements/requirements.in
django-imagekit==6.0.0
    # via -r requirements/requirements.in
django-rest-framework==0.1.0
    # via -r requirements/requirements.in
djangorestframework==3.16.1
    # via django-rest-framework
filelock==3.20.3
    # via tldextract
gevent==25.9.1
    # via -r requirements/requirements.in
googletrans==4.0.2
    # via -r requirements/requirements.in
greenlet==3.3.0
    # via gevent
gunicorn==23.0.0
    # via -r requirements/requirements.in
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via googletrans
hyperframe==6.1.0
    # via h2
hyperlink==21.0.0
    # via twisted
idna==3.11
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   tldextract
incremental==24.11.0
    # via twisted
itemadapter==0.13.1
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.3.2
    # via scrapy
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
lxml==6.0.2
    # via
    #   -r requirements/requirements.in
    #   parsel
    #   scrapy
ochre==0.4.0
    # via stransi
packaging==25.0
    # via
    #   gunicorn
    #   incremental
    #   parsel
    #   scrapy
parsel==1.10.0
    # via
    #   itemloaders
    #   scrapy
pilkit==3.0
    # via django-imagekit
pillow==12.1.0
    # via
    #   -r requirements/requirements.in
    #   django-colorfield
    #   pilkit
protego==0.5.0
    # via scrapy
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pydispatcher==2.0.7
    # via scrapy
pyopenssl==25.3.0
    # via scrapy
python-slugify==8.0.4
    # via django-admin-interface
queuelib==1.8.0
    # via scrapy
requests==2.32.5
    # via
    #   -r requirements/requirements.in
    #   requests-file
    #   tldextract
requests-file==3.0.1
    # via tldextract
scrapy==2.14.1
    # via -r requirements/requirements.in
service-identity==24.2.0
    # via scrapy
sqlparse==0.5.5
    # via
    #   django
    #   django-debug-toolbar
stransi==0.3.0
    # via -r requirements/requirements.in
telnetlib3==2.0.8
    # via -r requirements/requirements.in
text-unidecode==1.3
    # via python-slugify
tldextract==5.3.1
    # via scrapy
twisted==25.5.0
    # via scrapy
typing-extensions==4.15.0
    # via
    #   anyio
    #   pyopenssl
    #   twisted
urllib3==2.6.3
    # via requests
w3lib==2.3.1
    # via
    #   parsel
    #   scrapy
zope-event==6.1
    # via gevent
zope-interface==8.2
    # via
    #   gevent
    #   scrapy
    #   twisted
